requirements.txt
fastapi==0.65.2
psycopg2-binary==2.8.6
pydantic==1.8.2
python-dotenv==0.17.1
requests==2.26.0
SQLAlchemy==1.4.18
urllib3==1.26.6